+++ lib/noise/crypto/dh.ex
@@ -1,4 +1,10 @@
 defmodule Noise.Crypto.DH do
+  @moduledoc """
+  Behaviour for Diffie-Hellman functions.
+
+  This module defines the interface that all DH implementations must follow,
+  including key generation and performing the DH calculation.
+  """
   @type pubkey() :: <<_::_*8>>
   @type keypair() :: {<<_::_*8>>, pubkey()}