+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+- implemented PKS support
+- implemented defferred protocols support
+- added "public API" in the `Noise` module
+- added support for running tests from test vector files
+
 ## v0.1.0
 
 - initial implementation